fix: move request statuses to types file in ducks and remove styled in App component
